Ability to list content's of ships cargo hold.
Design decision was made to store cargo as a HashMap of Strings
mapped to a quantity, rather than any reference to a Commodity.
I think this will be more flexible and in any case the ship's
cargo hold should not need to know about commodities, just the
labels on their containers.