Switch from message passing to shared state concurrency.
The simulation thread updates the simulation state once per second.
The command handler in the main thread calls methods on the simulation
state to effect commands.
The simulation state can be accessed from both threads because it
access to it is controlled by a mutex. It is owned by both threads
because the mutex is wrapped in an atomic reference count pointer.